refactor: EmbeddingModel only implements embed_chunks
