+++ ragna/core/_rag.py
@@ -8,6 +8,7 @@ from typing import (
     AsyncIterator,
     Awaitable,
     Callable,
+    Deque,
     Generic,
     Iterable,
     Iterator,
@@ -18,11 +19,15 @@ from typing import (
     cast,
 )
 
+from collections import deque
+
 import pydantic
+import tiktoken
 from starlette.concurrency import iterate_in_threadpool, run_in_threadpool
 
 from ._components import (
     Assistant,
+    Chunk,
     Component,
     Embedding,
     EmbeddingModel,
@@ -30,8 +35,10 @@ from ._components import (
     MessageRole,
     SourceStorage,
 )
-from ._document import Document, LocalDocument
+from ._document import Document, LocalDocument, Page
 from ._utils import RagnaException, default_user, merge_models
+
+from uuid import UUID
 
 T = TypeVar("T")
 C = TypeVar("C", bound=Component)
@@ -118,6 +125,25 @@ class SpecialChatParams(pydantic.BaseModel):
     )
 
 
+# The function is adapted from more_itertools.windowed to allow a ragged last window
+# https://more-itertools.readthedocs.io/en/stable/api.html#more_itertools.windowed
+def windowed_ragged(
+    iterable: Iterable[T], *, n: int, step: int
+) -> Iterator[tuple[T, ...]]:
+    window: Deque[T] = deque(maxlen=n)
+    i = n
+    for _ in map(window.append, iterable):
+        i -= 1
+        if not i:
+            i = step
+            yield tuple(window)
+
+    if len(window) < n:
+        yield tuple(window)
+    elif 0 < i < min(step, n):
+        yield tuple(window)[i:]
+
+
 class Chat:
     """
     !!! tip
@@ -165,6 +191,8 @@ class Chat:
 
         self.documents = self._parse_documents(documents)
 
+        self._tokenizer = tiktoken.get_encoding('cl100k_base')
+
         if embedding_model is None and issubclass(
             source_storage.__ragna_input_type__, Embedding
         ):
@@ -189,6 +217,38 @@ class Chat:
         self._prepared = False
         self._messages: list[Message] = []
 
+    def _embed_text(self, text: str) -> list[float]:
+        dummy_chunk = Chunk(text=text, document_id=uuid.uuid4(), page_numbers=[], num_tokens=0)
+        return cast(EmbeddingModel, self.embedding_model).embed_chunks([dummy_chunk])[0].embedding
+
+    def _chunk_pages(
+        self,
+        pages: Iterable[Page],
+        document_id: UUID,
+        *,
+        chunk_size: int,
+        chunk_overlap: int,
+    ) -> Iterator[Chunk]:
+        self._tokenizer = tiktoken.get_encoding("cl100k_base")
+
+        for window in windowed_ragged(
+            (
+                (tokens, page.number)
+                for page in pages
+                for tokens in self._tokenizer.encode(page.text)
+            ),
+            n=chunk_size,
+            step=chunk_size - chunk_overlap,
+        ):
+            tokens, page_numbers = zip(*window)
+            yield Chunk(
+                text=self._tokenizer.decode(tokens),  # type: ignore[arg-type]
+                document_id=document_id,
+                page_numbers=list(filter(lambda n: n is not None, page_numbers))
+                or None,
+                num_tokens=len(tokens),
+            )
+
     async def prepare(self) -> Message:
         """Prepare the chat.
 
@@ -210,9 +270,16 @@ class Chat:
                 detail=RagnaException.EVENT,
             )
 
+        chunks = [chunk for document in self.documents for chunk in self._chunk_pages(
+                document.extract_pages(),
+                document_id=document.id,
+                chunk_size=500,
+                chunk_overlap=250,
+        )]
+
         input: Union[list[Document], list[Embedding]] = self.documents
         if not issubclass(self.source_storage.__ragna_input_type__, Document):
-            input = cast(EmbeddingModel, self.embedding_model).embed_documents(input)
+            input = cast(EmbeddingModel, self.embedding_model).embed_chunks(chunks)
         await self._run(self.source_storage.store, input)
 
         self._prepared = True
@@ -248,7 +315,7 @@ class Chat:
         if not issubclass(self.source_storage.__ragna_input_type__, Document):
             input = cast(
                 list[float],
-                cast(EmbeddingModel, self.embedding_model).embed_text(prompt),
+                self._embed_text(prompt),
             )
         sources = await self._run(self.source_storage.retrieve, self.documents, input)
